feat(scripts): add estimate size for download/installed system from config
sybling to `count` this allows for inspection before installing
