+++ archinstoo/archinstoo/scripts/list.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from archinstoo import DEFAULT, ROOTLESS_SCRIPTS
 
-scripts = [p.stem for p in Path(__file__).parent.glob('*.py') if p.stem not in ('__init__', 'list')]
+scripts = [p.stem for p in Path(__file__).parent.glob('*.py') if not p.stem.startswith('_') and p.stem != 'list']
 
 rootless = sorted(s for s in scripts if s in ROOTLESS_SCRIPTS)
 root = sorted(s for s in scripts if s not in ROOTLESS_SCRIPTS)
